ISSUES: RTK implementation errors and failure
current update on search bar
