+++ src/features/appBar/AppBar.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Stack, styled, AppBar as MuiAppBar, Toolbar } from '@mui/material';
-import SearchBar from './SearchBar';
+import SearchBar from '@/components/EntitledElement/SearchBar';
 
 const BarHeader = styled(Stack)(({ theme }) => ({
   color: theme.palette.common.white,
